strata-backend-engine/backend/routes/connections.py
from fastapi import APIRouter, HTTPException
from backend.models import ConnectionTestRequest, ConnectionTestResponse, ConnectionSaveRequest, ConnectionSaveResponse, ConnectionResponse
from backend.database import save_connection, get_all_connections, get_connection_by_id, update_connection
from typing import List
import importlib
import sqlite3

router = APIRouter()

def get_db_connector(db_type: str):
    """Dynamically import and return the appropriate database connector"""
    connectors = {
        "PostgreSQL": "psycopg2",
        "MySQL": "mysql.connector",
        "Snowflake": "snowflake.connector",
        "Databricks": "databricks.sql",
        "Oracle": "oracledb",
        "SQL Server": "pyodbc",
        "Teradata": "teradatasql",
        "Google BigQuery": "google.cloud.bigquery"
    }
    
    try:
        if db_type in connectors:
            return importlib.import_module(connectors[db_type])
        return None
    except ImportError:
        return None

def test_postgresql_connection(credentials):
    """Test PostgreSQL database connection with proper SSL handling for Azure"""
    try:
        # Import psycopg2 inside the function to handle import errors
        import psycopg2
        import psycopg2.extras
        
        # Extract credentials
        host = credentials.get('host')
        port = credentials.get('port', 5432)
        database = credentials.get('database')
        username = credentials.get('username')
        password = credentials.get('password')
        ssl_mode = credentials.get('sslmode', 'require')  # Default to require for PostgreSQL
        
        # Try different authentication approaches for Azure
        approaches = [
            # Approach 1: Standard username (without @hostname) - works for many Azure instances
            {"user": username},
            # Approach 2: Username with @hostname - required for some Azure instances
            {"user": f"{username}@{host.split('.')[0]}" if '.' in host else username},
        ]
        
        last_error = None
        
        for approach in approaches:
            try:
                # Configure connection parameters
                connection_params = {
                    'host': host,
                    'port': port,
                    'database': database,
                    'password': password,
                    'sslmode': ssl_mode,
                    **approach
                }
                
                # Create connection
                connection = psycopg2.connect(**connection_params)
                
                # Get server version
                cursor = connection.cursor()
                cursor.execute("SELECT version();")
                version = cursor.fetchone()[0]
                cursor.close()
                connection.close()
                
                return True, version
            except Exception as e:
                last_error = e
                continue
        
        # If we get here, all approaches failed
        return False, f"Connection failed: {last_error}"
        
    except ImportError:
        return False, "PostgreSQL connector (psycopg2) not installed"
    except Exception as e:
        # Handle any psycopg2 specific exceptions
        error_msg = str(e)
        if "psycopg2" in error_msg.lower() or "postgresql" in error_msg.lower():
            return False, error_msg
        return False, f"Connection failed: {error_msg}"

def test_mysql_connection(credentials):
    """Test MySQL database connection with proper SSL handling for Azure"""
    try:
        # Import mysql.connector inside the function to handle import errors
        import mysql.connector
        
        # Extract credentials
        host = credentials.get('host')
        port = credentials.get('port', 3306)
        database = credentials.get('database')
        username = credentials.get('username')
        password = credentials.get('password')
        ssl_mode = credentials.get('ssl', 'true')  # Default to true for MySQL
        
        # Configure SSL settings based on the sslmode
        ssl_config = {}
        if ssl_mode == 'false':
            ssl_config['ssl_disabled'] = True
        else:
            # For Azure MySQL, we need to handle SSL properly
            ssl_config['ssl_disabled'] = False
            # Don't verify certificate for Azure MySQL unless specifically configured
            ssl_config['ssl_verify_cert'] = False
            ssl_config['ssl_verify_identity'] = False
        
        # Create connection with SSL configuration
        connection_params = {
            'host': host,
            'port': port,
            'database': database,
            'user': username,
            'password': password,
            **ssl_config
        }
        
        connection = mysql.connector.connect(**connection_params)
        
        if connection.is_connected():
            db_info = connection.get_server_info()
            connection.close()
            return True, db_info
        else:
            return False, "Failed to connect"
    except ImportError:
        return False, "MySQL connector not installed"
    except Exception as e:
        # Handle any mysql.connector specific exceptions
        error_msg = str(e)
        if "mysql" in error_msg.lower() or "connector" in error_msg.lower():
            return False, error_msg
        return False, f"Connection failed: {error_msg}"

def test_connection_by_type(db_type: str, credentials: dict):
    """Test connection based on database type"""
    try:
        # Check if we have the required connector
        connector = get_db_connector(db_type)
        
        if connector is None:
            return False, f"Driver not available for {db_type}. Using stub mode."
        
        # Actually test the connection based on database type
        if db_type == "PostgreSQL":
            return test_postgresql_connection(credentials)
        elif db_type == "MySQL":
            return test_mysql_connection(credentials)
        
        # For other database types, we would implement similar testing
        # For now, we'll just simulate a successful connection for non-PostgreSQL/MySQL databases
        return True, "8.0.0"
    except Exception as e:
        return False, str(e)

@router.post("/test", response_model=ConnectionTestResponse)
async def test_connection(request: ConnectionTestRequest):
    try:
        success, details = test_connection_by_type(request.dbType, request.credentials)
        
        if success:
            return ConnectionTestResponse(
                ok=True,
                vendorVersion=details,
                details=f"Successfully connected to {request.dbType}"
            )
        else:
            return ConnectionTestResponse(
                ok=False,
                details=f"Connection failed: {details}"
            )
    except Exception as e:
        return ConnectionTestResponse(
            ok=False,
            details=str(e)
        )

@router.post("/save", response_model=ConnectionSaveResponse)
async def save_connection_endpoint(request: ConnectionSaveRequest):
    try:
        connection_id = save_connection(
            name=request.name,
            db_type=request.dbType,
            credentials=request.credentials
        )
        
        return ConnectionSaveResponse(ok=True, id=connection_id)
    except Exception as e:
        return ConnectionSaveResponse(ok=False)

@router.put("/{connection_id}", response_model=ConnectionSaveResponse)
async def update_connection_endpoint(connection_id: int, request: ConnectionSaveRequest):
    try:
        success = update_connection(
            connection_id=connection_id,
            name=request.name,
            db_type=request.dbType,
            credentials=request.credentials
        )
        
        if success:
            return ConnectionSaveResponse(ok=True, id=connection_id)
        else:
            return ConnectionSaveResponse(ok=False)
    except Exception as e:
        return ConnectionSaveResponse(ok=False)

@router.get("/", response_model=List[ConnectionResponse])
async def list_connections():
    try:
        connections = get_all_connections()
        return connections
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@router.get("/{connection_id}", response_model=ConnectionResponse)
async def get_connection(connection_id: int):
    try:
        connection = get_connection_by_id(connection_id)
        if connection is None:
            raise HTTPException(status_code=404, detail="Connection not found")
        return connection
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@router.delete("/{connection_id}")
async def delete_connection(connection_id: int):
    """Delete a connection by ID"""
    try:
        conn = sqlite3.connect("strata.db")
        cursor = conn.cursor()
        
        cursor.execute("DELETE FROM connections WHERE id = ?", (connection_id,))
        conn.commit()
        conn.close()
        
        return {"ok": True, "message": "Connection deleted successfully"}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))