strata-backend-engine/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
cryptography==41.0.7
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
snowflake-connector-python==3.7.1
databricks-sql-connector==2.9.3
oracledb==1.4.2
google-cloud-bigquery==3.13.0
xlsxwriter==3.1.9
reportlab==4.0.7
python-multipart==0.0.6